🔧 chore #21: 스토리북 preview에 GlobalStyles 추가
